Fixed sql tx close, fixed unit tests. Refactoring. Added missing structs and methods.
