+++ merkletree.go
@@ -2,12 +2,10 @@ package merkletree
 
 import (
 	"bytes"
-	"encoding/hex"
 	"errors"
 	"fmt"
 	"io"
 	"math/big"
-	"strings"
 	"sync"
 
 	cryptoUtils "github.com/iden3/go-iden3-crypto/utils"
@@ -17,10 +15,13 @@ const (
 	// proofFlagsLen is the byte length of the flags in the proof header
 	// (first 32 bytes).
 	proofFlagsLen = 2
-	// ElemBytesLen is the length of the Hash byte array
-	ElemBytesLen = 32
 
 	numCharPrint = 8
+
+	// IndexLen indicates how many elements are used for the index.
+	IndexLen = 4
+	// DataLen indicates how many elements are used for the data.
+	DataLen = 8
 )
 
 var (
@@ -51,114 +52,7 @@ var (
 	ErrNotWritable = errors.New("Merkle Tree not writable")
 
 	dbKeyRootNode = []byte("currentroot")
-	// HashZero is used at Empty nodes
-	HashZero = Hash{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-		0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0}
 )
-
-// Hash is the generic type stored in the MerkleTree
-type Hash [32]byte
-
-// MarshalText implements the marshaler for the Hash type
-func (h Hash) MarshalText() ([]byte, error) {
-	return []byte(h.BigInt().String()), nil
-}
-
-// UnmarshalText implements the unmarshaler for the Hash type
-func (h *Hash) UnmarshalText(b []byte) error {
-	ha, err := NewHashFromString(string(b))
-	copy(h[:], ha[:])
-	return err
-}
-
-// String returns decimal representation in string format of the Hash
-func (h Hash) String() string {
-	s := h.BigInt().String()
-	if len(s) < numCharPrint {
-		return s
-	}
-	return s[0:numCharPrint] + "..."
-}
-
-// Hex returns the hexadecimal representation of the Hash
-func (h Hash) Hex() string {
-	return hex.EncodeToString(h[:])
-	// alternatively equivalent, but with too extra steps:
-	// bRaw := h.BigInt().Bytes()
-	// b := [32]byte{}
-	// copy(b[:], SwapEndianness(bRaw[:]))
-	// return hex.EncodeToString(b[:])
-}
-
-// BigInt returns the *big.Int representation of the *Hash
-func (h *Hash) BigInt() *big.Int {
-	if new(big.Int).SetBytes(SwapEndianness(h[:])) == nil {
-		return big.NewInt(0)
-	}
-	return new(big.Int).SetBytes(SwapEndianness(h[:]))
-}
-
-// Bytes returns the []byte representation of the *Hash, which always is 32
-// bytes length.
-func (h *Hash) Bytes() []byte {
-	bi := new(big.Int).SetBytes(h[:]).Bytes()
-	b := [32]byte{}
-	copy(b[:], SwapEndianness(bi[:]))
-	return b[:]
-}
-
-// NewBigIntFromHashBytes returns a *big.Int from a byte array, swapping the
-// endianness in the process. This is the intended method to get a *big.Int
-// from a byte array that previously has ben generated by the Hash.Bytes()
-// method.
-func NewBigIntFromHashBytes(b []byte) (*big.Int, error) {
-	if len(b) != ElemBytesLen {
-		return nil, fmt.Errorf("Expected 32 bytes, found %d bytes", len(b))
-	}
-	bi := new(big.Int).SetBytes(b[:ElemBytesLen])
-	if !cryptoUtils.CheckBigIntInField(bi) {
-		return nil, fmt.Errorf("NewBigIntFromHashBytes: Value not inside the Finite Field")
-	}
-	return bi, nil
-}
-
-// NewHashFromBigInt returns a *Hash representation of the given *big.Int
-func NewHashFromBigInt(b *big.Int) *Hash {
-	r := &Hash{}
-	copy(r[:], SwapEndianness(b.Bytes()))
-	return r
-}
-
-// NewHashFromBytes returns a *Hash from a byte array, swapping the endianness
-// in the process. This is the intended method to get a *Hash from a byte array
-// that previously has ben generated by the Hash.Bytes() method.
-func NewHashFromBytes(b []byte) (*Hash, error) {
-	if len(b) != ElemBytesLen {
-		return nil, fmt.Errorf("Expected 32 bytes, found %d bytes", len(b))
-	}
-	var h Hash
-	copy(h[:], SwapEndianness(b))
-	return &h, nil
-}
-
-// NewHashFromHex returns a *Hash representation of the given hex string
-func NewHashFromHex(h string) (*Hash, error) {
-	h = strings.TrimPrefix(h, "0x")
-	b, err := hex.DecodeString(h)
-	if err != nil {
-		return nil, err
-	}
-	return NewHashFromBytes(SwapEndianness(b[:]))
-}
-
-// NewHashFromString returns a *Hash representation of the given decimal string
-func NewHashFromString(s string) (*Hash, error) {
-	bi, ok := new(big.Int).SetString(s, 10)
-	if !ok {
-		return nil, fmt.Errorf("Can not parse string to Hash")
-	}
-	return NewHashFromBigInt(bi), nil
-}
 
 // MerkleTree is the struct with the main elements of the MerkleTree
 type MerkleTree struct {
@@ -258,6 +152,51 @@ func (mt *MerkleTree) Add(k, v *big.Int) error {
 	vHash := NewHashFromBigInt(v)
 	newNodeLeaf := NewNodeLeaf(kHash, vHash)
 	path := getPath(mt.maxLevels, kHash[:])
+
+	newRootKey, err := mt.addLeaf(tx, newNodeLeaf, mt.rootKey, 0, path)
+	if err != nil {
+		return err
+	}
+	mt.rootKey = newRootKey
+	err = mt.setCurrentRoot(tx, mt.rootKey)
+	if err != nil {
+		return err
+	}
+
+	if err := tx.Commit(); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+// AddEntry adds the Entry to the MerkleTree
+func (mt *MerkleTree) AddEntry(e *Entry) error {
+	// verify that the MerkleTree is writable
+	if !mt.writable {
+		return ErrNotWritable
+	}
+	// verify that the ElemBytes are valid and fit inside the mimc7 field.
+	if !CheckEntryInField(*e) {
+		return errors.New("Elements not inside the Finite Field over R")
+	}
+	tx, err := mt.db.NewTx()
+	if err != nil {
+		return err
+	}
+	mt.Lock()
+	defer mt.Unlock()
+
+	hIndex, err := e.HIndex()
+	if err != nil {
+		return err
+	}
+	hValue, err := e.HValue()
+	if err != nil {
+		return err
+	}
+	newNodeLeaf := NewNodeLeaf(hIndex, hValue)
+	path := getPath(mt.maxLevels, hIndex[:])
 
 	newRootKey, err := mt.addLeaf(tx, newNodeLeaf, mt.rootKey, 0, path)
 	if err != nil {
@@ -757,103 +696,6 @@ type NodeAux struct {
 	Value *Hash
 }
 
-// Proof defines the required elements for a MT proof of existence or
-// non-existence.
-type Proof struct {
-	// existence indicates wether this is a proof of existence or
-	// non-existence.
-	Existence bool
-	// depth indicates how deep in the tree the proof goes.
-	depth uint
-	// notempties is a bitmap of non-empty Siblings found in Siblings.
-	notempties [ElemBytesLen - proofFlagsLen]byte
-	// Siblings is a list of non-empty sibling keys.
-	Siblings []*Hash
-	NodeAux  *NodeAux
-}
-
-// NewProofFromBytes parses a byte array into a Proof.
-func NewProofFromBytes(bs []byte) (*Proof, error) {
-	if len(bs) < ElemBytesLen {
-		return nil, ErrInvalidProofBytes
-	}
-	p := &Proof{}
-	if (bs[0] & 0x01) == 0 {
-		p.Existence = true
-	}
-	p.depth = uint(bs[1])
-	copy(p.notempties[:], bs[proofFlagsLen:ElemBytesLen])
-	siblingBytes := bs[ElemBytesLen:]
-	sibIdx := 0
-	for i := uint(0); i < p.depth; i++ {
-		if TestBitBigEndian(p.notempties[:], i) {
-			if len(siblingBytes) < (sibIdx+1)*ElemBytesLen {
-				return nil, ErrInvalidProofBytes
-			}
-			var sib Hash
-			copy(sib[:], siblingBytes[sibIdx*ElemBytesLen:(sibIdx+1)*ElemBytesLen])
-			p.Siblings = append(p.Siblings, &sib)
-			sibIdx++
-		}
-	}
-
-	if !p.Existence && ((bs[0] & 0x02) != 0) {
-		p.NodeAux = &NodeAux{Key: &Hash{}, Value: &Hash{}}
-		nodeAuxBytes := siblingBytes[len(p.Siblings)*ElemBytesLen:]
-		if len(nodeAuxBytes) != 2*ElemBytesLen {
-			return nil, ErrInvalidProofBytes
-		}
-		copy(p.NodeAux.Key[:], nodeAuxBytes[:ElemBytesLen])
-		copy(p.NodeAux.Value[:], nodeAuxBytes[ElemBytesLen:2*ElemBytesLen])
-	}
-	return p, nil
-}
-
-// Bytes serializes a Proof into a byte array.
-func (p *Proof) Bytes() []byte {
-	bsLen := proofFlagsLen + len(p.notempties) + ElemBytesLen*len(p.Siblings)
-	if p.NodeAux != nil {
-		bsLen += 2 * ElemBytesLen //nolint:gomnd
-	}
-	bs := make([]byte, bsLen)
-
-	if !p.Existence {
-		bs[0] |= 0x01
-	}
-	bs[1] = byte(p.depth)
-	copy(bs[proofFlagsLen:len(p.notempties)+proofFlagsLen], p.notempties[:])
-	siblingsBytes := bs[len(p.notempties)+proofFlagsLen:]
-	for i, k := range p.Siblings {
-		copy(siblingsBytes[i*ElemBytesLen:(i+1)*ElemBytesLen], k[:])
-	}
-	if p.NodeAux != nil {
-		bs[0] |= 0x02
-		copy(bs[len(bs)-2*ElemBytesLen:], p.NodeAux.Key[:])
-		copy(bs[len(bs)-1*ElemBytesLen:], p.NodeAux.Value[:])
-	}
-	return bs
-}
-
-// SiblingsFromProof returns all the siblings of the proof.
-func SiblingsFromProof(proof *Proof) []*Hash {
-	sibIdx := 0
-	siblings := []*Hash{}
-	for lvl := 0; lvl < int(proof.depth); lvl++ {
-		if TestBitBigEndian(proof.notempties[:], uint(lvl)) {
-			siblings = append(siblings, proof.Siblings[sibIdx])
-			sibIdx++
-		} else {
-			siblings = append(siblings, &HashZero)
-		}
-	}
-	return siblings
-}
-
-// AllSiblings returns all the siblings of the proof.
-func (p *Proof) AllSiblings() []*Hash {
-	return SiblingsFromProof(p)
-}
-
 // CircomSiblingsFromSiblings returns the full siblings compatible with circom
 func CircomSiblingsFromSiblings(siblings []*Hash, levels int) []*Hash {
 	// Add the rest of empty levels to the siblings
@@ -1008,67 +850,6 @@ func (mt *MerkleTree) GenerateProof(k *big.Int, rootKey *Hash) (*Proof,
 	return nil, nil, ErrKeyNotFound
 }
 
-// VerifyProof verifies the Merkle Proof for the entry and root.
-func VerifyProof(rootKey *Hash, proof *Proof, k, v *big.Int) bool {
-	rootFromProof, err := RootFromProof(proof, k, v)
-	if err != nil {
-		return false
-	}
-	return bytes.Equal(rootKey[:], rootFromProof[:])
-}
-
-// RootFromProof calculates the root that would correspond to a tree whose
-// siblings are the ones in the proof with the leaf hashing to hIndex and
-// hValue.
-func RootFromProof(proof *Proof, k, v *big.Int) (*Hash, error) {
-	kHash := NewHashFromBigInt(k)
-	vHash := NewHashFromBigInt(v)
-	sibIdx := len(proof.Siblings) - 1
-	var err error
-	var midKey *Hash
-	if proof.Existence {
-		midKey, err = LeafKey(kHash, vHash)
-		if err != nil {
-			return nil, err
-		}
-	} else {
-		if proof.NodeAux == nil {
-			midKey = &HashZero
-		} else {
-			if bytes.Equal(kHash[:], proof.NodeAux.Key[:]) {
-				return nil,
-					fmt.Errorf("Non-existence proof being checked against hIndex equal to nodeAux")
-			}
-			midKey, err = LeafKey(proof.NodeAux.Key, proof.NodeAux.Value)
-			if err != nil {
-				return nil, err
-			}
-		}
-	}
-	path := getPath(int(proof.depth), kHash[:])
-	var siblingKey *Hash
-	for lvl := int(proof.depth) - 1; lvl >= 0; lvl-- {
-		if TestBitBigEndian(proof.notempties[:], uint(lvl)) {
-			siblingKey = proof.Siblings[sibIdx]
-			sibIdx--
-		} else {
-			siblingKey = &HashZero
-		}
-		if path[lvl] {
-			midKey, err = NewNodeMiddle(siblingKey, midKey).Key()
-			if err != nil {
-				return nil, err
-			}
-		} else {
-			midKey, err = NewNodeMiddle(midKey, siblingKey).Key()
-			if err != nil {
-				return nil, err
-			}
-		}
-	}
-	return midKey, nil
-}
-
 // walk is a helper recursive function to iterate over all tree branches
 func (mt *MerkleTree) walk(key *Hash, f func(*Node)) error {
 	n, err := mt.GetNode(key)
@@ -1199,3 +980,37 @@ func (mt *MerkleTree) ImportDumpedLeafs(b []byte) error {
 	}
 	return nil
 }
+
+//// ImportTree imports the tree from the output from the DumpTree function
+//func (mt *MerkleTree) ImportTree(i io.Reader) error {
+//	tx, err := mt.DB().NewTx()
+//	if err != nil {
+//		return err
+//	}
+//	mt.Lock()
+//	defer mt.Unlock()
+//
+//	r := bufio.NewReader(i)
+//	for {
+//		k, v, err := deserializeKV(r)
+//		if err == io.EOF {
+//			break
+//		} else if err != nil {
+//			return err
+//		}
+//		tx.Put(k, v)
+//	}
+//
+//	v, err := tx.GetRoot()
+//	if err != nil {
+//		return err
+//	}
+//
+//	if err := tx.Commit(); err != nil {
+//		return err
+//	}
+//	mt.rootKey = &Hash{}
+//	copy(mt.rootKey[:], v[:])
+//
+//	return nil
+//}
